Extract CSS module
Extracts CSS specific functions from `src/dom.js` into `src/css.js`.
