+++ build-system/tasks/presubmit-checks.js
@@ -816,12 +816,6 @@ const forbiddenTermsSrcInclusive = {
       'extensions/amp-bind/0.1/bind-expr-impl.js',
     ],
   },
-  'scopeSelectorForTesting': {
-    message: 'scopeSelector is not intended to be used outside of dom.js',
-    whitelist: [
-      'src/dom.js',
-    ],
-  },
   '[^.]loadPromise': {
     message: 'Most users should use BaseElement…loadPromise.',
     whitelist: [
