+++ extensions/amp-access/0.1/amp-access-server.js
@@ -18,7 +18,7 @@ import {AccessClientAdapter} from './amp-access-client';
 import {Services} from '../../../src/services';
 import {dev, devAssert} from '../../../src/log';
 import {dict} from '../../../src/utils/object';
-import {escapeCssSelectorIdent} from '../../../src/dom';
+import {escapeCssSelectorIdent} from '../../../src/css';
 import {fetchDocument} from '../../../src/document-fetcher';
 import {isExperimentOn} from '../../../src/experiments';
 import {isProxyOrigin, removeFragment} from '../../../src/url';
