+++ extensions/amp-form/0.1/amp-form.js
@@ -37,7 +37,6 @@ import {
   ancestorElementsByTag,
   childElementByAttr,
   createElementWithAttributes,
-  escapeCssSelectorIdent,
   iterateCursor,
   removeElement,
   tryFocus,
@@ -46,6 +45,7 @@ import {createCustomEvent} from '../../../src/event-helper';
 import {createFormDataWrapper} from '../../../src/form-data-wrapper';
 import {deepMerge, dict} from '../../../src/utils/object';
 import {dev, devAssert, user, userAssert} from '../../../src/log';
+import {escapeCssSelectorIdent} from '../../../src/css';
 import {
   formOrNullForElement,
   getFormAsObject,
