+++ extensions/amp-story/0.1/page-advancement.js
@@ -15,8 +15,9 @@
  */
 import {Services} from '../../../src/services';
 import {VideoEvents} from '../../../src/video-interface';
-import {closest, escapeCssSelectorIdent} from '../../../src/dom';
+import {closest} from '../../../src/dom';
 import {dev, user} from '../../../src/log';
+import {escapeCssSelectorIdent} from '../../../src/css';
 import {hasTapAction, timeStrToMillis} from './utils';
 import {listenOnce} from '../../../src/event-helper';
 import {map} from '../../../src/utils/object';
