+++ extensions/amp-story/0.1/progress-bar.js
@@ -16,9 +16,10 @@
 import {POLL_INTERVAL_MS} from './page-advancement';
 import {Services} from '../../../src/services';
 import {dev, devAssert} from '../../../src/log';
-import {escapeCssSelectorNth, scopedQuerySelector} from '../../../src/dom';
+import {escapeCssSelectorNth} from '../../../src/css';
 import {hasOwn, map} from '../../../src/utils/object';
 import {scale, setImportantStyles} from '../../../src/style';
+import {scopedQuerySelector} from '../../../src/dom';
 
 
 /**
