+++ extensions/amp-story/1.0/animation.js
@@ -28,13 +28,13 @@ import {
 } from '../../amp-animation/0.1/web-animation-types';
 import {assertDoesNotContainDisplay, setStyles} from '../../../src/style';
 import {dev, devAssert, user, userAssert} from '../../../src/log';
+import {escapeCssSelectorIdent} from '../../../src/css';
+import {getPresetDef, setStyleForPreset} from './animation-presets';
+import {map, omit} from '../../../src/utils/object';
 import {
-  escapeCssSelectorIdent,
   scopedQuerySelector,
   scopedQuerySelectorAll,
 } from '../../../src/dom';
-import {getPresetDef, setStyleForPreset} from './animation-presets';
-import {map, omit} from '../../../src/utils/object';
 import {timeStrToMillis, unscaledClientRect} from './utils';
 
 /** const {string} */
