+++ extensions/amp-web-push/0.1/amp-web-push-config.js
@@ -16,7 +16,7 @@
 
 import {CONFIG_TAG, SERVICE_TAG, TAG} from './vars';
 import {dev, user, userAssert} from '../../../src/log';
-import {escapeCssSelectorIdent} from '../../../src/dom';
+import {escapeCssSelectorIdent} from '../../../src/css';
 import {getServiceForDoc} from '../../../src/service';
 import {parseUrlDeprecated} from '../../../src/url';
