+++ extensions/amp-web-push/0.1/amp-web-push-permission-dialog.js
@@ -16,7 +16,7 @@
 
 import {NotificationPermission, StorageKeys} from './vars';
 import {WindowMessenger} from './window-messenger';
-import {escapeCssSelectorIdent} from '../../../src/dom';
+import {escapeCssSelectorIdent} from '../../../src/css';
 import {getMode} from '../../../src/mode';
 import {
   parseQueryString,
