+++ src/dom.js
@@ -15,9 +15,9 @@
  */
 
 import {Deferred} from './utils/promise';
-import {cssEscape} from '../third_party/css-escape/css-escape';
 import {dev, devAssert} from './log';
 import {dict} from './utils/object';
+import {isScopeSelectorSupported, prependSelectorsWith} from './css';
 import {startsWith} from './string';
 import {toWin} from './types';
 
@@ -481,60 +481,6 @@ function assertIsTagName(tagName) {
 }
 
 /**
- * @type {boolean|undefined}
- * @visibleForTesting
- */
-let scopeSelectorSupported;
-
-/**
- * @param {boolean|undefined} val
- * @visibleForTesting
- */
-export function setScopeSelectorSupportedForTesting(val) {
-  scopeSelectorSupported = val;
-}
-
-/**
- * Test that the :scope selector is supported and behaves correctly.
- * @param {!Element} parent
- * @return {boolean}
- */
-function isScopeSelectorSupported(parent) {
-  const doc = parent.ownerDocument;
-  try {
-    const testElement = doc.createElement('div');
-    const testChild = doc.createElement('div');
-    testElement.appendChild(testChild);
-    // NOTE(cvializ, #12383): Firefox's implementation is incomplete,
-    // therefore we test actual functionality of`:scope` as well.
-    return testElement./*OK*/querySelector(':scope div') === testChild;
-  } catch (e) {
-    return false;
-  }
-}
-
-/**
- * Prefixes a selector for ancestor selection. Splits in subselectors and
- * applies prefix to each.
- *
- * e.g.
- * ```
- *   scopeSelector('.i-amphtml-scoped', 'div'); // .i-amphtml-scoped div
- *   scopeSelector(':scope', 'div, ul');        // :scope div, :scope ul
- *   scopeSelector('article >', 'div, ul');     // article > div, article > ul
- * ```
- *
- * @param {string} distribute
- * @param {string} selector
- * @return {string}
- */
-function scopeSelector(distribute, selector) {
-  return selector.replace(/^|,/g, `$&${distribute} `);
-}
-
-export const scopeSelectorForTesting = scopeSelector;
-
-/**
  * Finds all elements that matche `selector`, scoped inside `root`
  * for user-agents that do not support native scoping.
  *
@@ -547,7 +493,7 @@ export const scopeSelectorForTesting = scopeSelector;
 function scopedQuerySelectionFallback(root, selector) {
   const unique = 'i-amphtml-scoped';
   root.classList.add(unique);
-  const scopedSelector = scopeSelector(`.${unique}`, selector);
+  const scopedSelector = prependSelectorsWith(selector, `.${unique}`);
   const elements = root./*OK*/querySelectorAll(scopedSelector);
   root.classList.remove(unique);
   return elements;
@@ -561,11 +507,8 @@ function scopedQuerySelectionFallback(root, selector) {
  * @return {?Element}
  */
 export function scopedQuerySelector(root, selector) {
-  if (scopeSelectorSupported == null) {
-    scopeSelectorSupported = isScopeSelectorSupported(root);
-  }
-  if (scopeSelectorSupported) {
-    return root./*OK*/querySelector(scopeSelector(':scope', selector));
+  if (isScopeSelectorSupported(root)) {
+    return root./*OK*/querySelector(prependSelectorsWith(selector, ':scope'));
   }
 
   // Only IE.
@@ -581,11 +524,9 @@ export function scopedQuerySelector(root, selector) {
  * @return {!NodeList<!Element>}
  */
 export function scopedQuerySelectorAll(root, selector) {
-  if (scopeSelectorSupported == null) {
-    scopeSelectorSupported = isScopeSelectorSupported(root);
-  }
-  if (scopeSelectorSupported) {
-    return root./*OK*/querySelectorAll(scopeSelector(':scope', selector));
+  if (isScopeSelectorSupported(root)) {
+    return root./*OK*/querySelectorAll(
+        prependSelectorsWith(selector, ':scope'));
   }
 
   // Only IE.
@@ -733,34 +674,6 @@ export function isRTL(doc) {
                  || doc.documentElement.getAttribute('dir')
                  || 'ltr';
   return dir == 'rtl';
-}
-
-
-/**
- * Escapes an ident (ID or a class name) to be used as a CSS selector.
- *
- * See https://drafts.csswg.org/cssom/#serialize-an-identifier.
- *
- * @param {string} ident
- * @return {string}
- */
-export function escapeCssSelectorIdent(ident) {
-  return cssEscape(ident);
-}
-
-/**
- * Escapes an ident in a way that can be used by :nth-child() psuedo-class.
- *
- * See https://github.com/w3c/csswg-drafts/issues/2306.
- *
- * @param {string|number} ident
- * @return {string}
- */
-export function escapeCssSelectorNth(ident) {
-  const escaped = String(ident);
-  // Ensure it doesn't close the nth-child psuedo class.
-  devAssert(escaped.indexOf(')') === -1);
-  return escaped;
 }
 
 /**
