+++ test/unit/test-dom.js
@@ -18,6 +18,7 @@ import * as dom from '../../src/dom';
 import {BaseElement} from '../../src/base-element';
 import {createAmpElementForTesting} from '../../src/custom-element';
 import {loadPromise} from '../../src/event-helper';
+import {setScopeSelectorSupportedForTesting} from '../../src/css';
 import {toArray} from '../../src/types';
 
 
@@ -30,29 +31,8 @@ describes.sandboxed('DOM', {}, env => {
   });
 
   afterEach(() => {
-    dom.setScopeSelectorSupportedForTesting(undefined);
+    setScopeSelectorSupportedForTesting(undefined);
     sandbox.restore();
-  });
-
-  describe('scopeSelector', () => {
-
-    const {scopeSelectorForTesting: scopeSelector} = dom;
-
-    it('concats simple', () => {
-      expect(scopeSelector('.i-amphtml-scoped', 'div'))
-          .to.equal('.i-amphtml-scoped div');
-    });
-
-    it('concats multiple selectors (2)', () => {
-      expect(scopeSelector(':scope', 'div,ul'))
-          .to.equal(':scope div,:scope ul');
-    });
-
-    it('concats multiple selectors (4)', () => {
-      expect(scopeSelector('div >', 'div,ul,ol,section'))
-          .to.equal('div > div,div > ul,div > ol,div > section');
-    });
-
   });
 
   it('should remove all children', () => {
@@ -378,7 +358,7 @@ describes.sandboxed('DOM', {}, env => {
   it('childElementByTag should find first match', testChildElementByTag);
 
   it('childElementByTag should find first match (polyfill)', () => {
-    dom.setScopeSelectorSupportedForTesting(false);
+    setScopeSelectorSupportedForTesting(false);
     testChildElementByTag();
   });
 
@@ -405,7 +385,7 @@ describes.sandboxed('DOM', {}, env => {
   it('childElementsByTag should find first match', testChildElementsByTag);
 
   it('childElementsByTag should find first match (polyfill)', () => {
-    dom.setScopeSelectorSupportedForTesting(false);
+    setScopeSelectorSupportedForTesting(false);
     testChildElementsByTag();
   });
 
@@ -436,7 +416,7 @@ describes.sandboxed('DOM', {}, env => {
   it('childElementByAttr should find first match', testChildElementByAttr);
 
   it('childElementByAttr should find first match', () => {
-    dom.setScopeSelectorSupportedForTesting(false);
+    setScopeSelectorSupportedForTesting(false);
     testChildElementByAttr();
   });
 
@@ -467,7 +447,7 @@ describes.sandboxed('DOM', {}, env => {
   it('childElementsByAttr should find all matches', testChildElementsByAttr);
 
   it('childElementsByAttr should find all matches', () => {
-    dom.setScopeSelectorSupportedForTesting(false);
+    setScopeSelectorSupportedForTesting(false);
     testChildElementsByAttr();
   });
 
@@ -561,7 +541,7 @@ describes.sandboxed('DOM', {}, env => {
   it('scopedQuerySelector should find first match', testScopedQuerySelector);
 
   it('scopedQuerySelector should find first match (polyfill)', () => {
-    dom.setScopeSelectorSupportedForTesting(false);
+    setScopeSelectorSupportedForTesting(false);
     testScopedQuerySelector();
   });
 
@@ -588,7 +568,7 @@ describes.sandboxed('DOM', {}, env => {
       testScopedQuerySelectorAll);
 
   it('scopedQuerySelectorAll should find all matches (polyfill)', () => {
-    dom.setScopeSelectorSupportedForTesting(false);
+    setScopeSelectorSupportedForTesting(false);
     testScopedQuerySelectorAll();
   });
